created a FlopMeasure class on top of FlopCounterMode instead of keep of copy of our own tflop_counter.py
